feat/ASF-15: изменил логику в select
