--- src/components/content-select/contentSelect.constans.ts
@@ -1,4 +1,0 @@
-export const menuSelectItem: string[] = [
-  'Жанры', 'Фантастика', 'Фэнтези',
-  'Боевик', 'Триллер', 'Ужасы'
-]
