+++ src/components/filter/Filter.tsx
@@ -1,7 +1,6 @@
 import style from './styles.module.scss'
 import cnBind from 'classnames/bind'
 import { Select } from '@/ui'
-import { ContentSelect } from '@/components'
 import { useState, MouseEvent } from 'react'
 
 const cx = cnBind.bind(style)
@@ -25,9 +24,10 @@ export const Filter = () => {
         <div className={cx('filter__body')}>
           <div className={cx('filter__item')}>
             <Select
-              option={<ContentSelect checkIconPosition={'right'} onClick={onValueChange}/>}
+              data={[{ value: 'value', label: 'label' }]}
               value={a}
               onClearValue={onClearValue}
+              onChange={(value) => console.log(value)}
             />
           </div>
         </div>
