+++ src/components/index.ts
@@ -11,5 +11,3 @@ export { Navbar } from '@components/navbar/Navbar'
 export { SendNewPasswordForm } from '@components/send-email/SendNewPasswordForm'
 
 export { Filter } from '@components/filter/Filter'
-
-export { ContentSelect } from '@components/content-select/ContentSelect'
