+++ src/types/common.ts
@@ -1,0 +1,4 @@
+export type OptionType = {
+  value: string
+  label: string
+}
